requirements.txt
asgiref
autopep8
Django==3.1.*
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
djangorestframework